+++ rootdriver/agent.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from uuid import uuid4
-
+import asyncio
 from .engine import Engine
 from .llm import LLM
 from .conversation import Conversation
@@ -19,7 +19,7 @@ class Agent:
         system_prompt: str | None = None,
 
         db_path: str| None = None,
-        # retry: int = 3,
+        # llm_retry: int = 3,
         # timeout: float | None = None,
     ):
         self.id = id if id else uuid4().hex
@@ -45,6 +45,17 @@ class Agent:
     def talk(self, input_prompt:str) -> "str":
         """一次 agent 调用"""
         return self.engine.invoke(build_user_message(input_prompt)).content
+
+
+    """==========异步部分=========="""
+    async def areact(self, input_prompt:str) -> "str":
+        """一次 react 异步循环"""
+        response = await self.engine.arun(build_user_message(input_prompt))
+        return response.content
+
+    async def atalk(self, input_prompt:str) -> "str":
+        """一次 agent 异步调用"""
+        return await self.engine.ainvoke(build_user_message(input_prompt)).content