+++ rootdriver/engine.py
@@ -1,4 +1,4 @@
-
+import asyncio
 
 from .llm import LLM
 from .conversation import Conversation
@@ -14,18 +14,16 @@ class Engine:
         self.model = model
 
     def invoke(self, input_message:Message) -> Message:
+        """但系调用llm，并存入记忆，无tool调用"""
         self.conversation.append(input_message)
         message = self.chat()
         return message
 
     def chat(self) -> Message:
         """
-
         1. 读会话消息
         2. 大模型交流
         3. 追加会话消息
-        4. 判断 是否调用工具
-        5. 是： tool 返回结果并追加，否： 追加并返回
         """
         # 1. 读会话消息
         message = self.conversation.get_messages()
@@ -41,17 +39,9 @@ class Engine:
 
     def deal_tool_or_output(self, response_message:Message) -> Message|None:
         """
-                1. 追加输入
-                2. 读会话消息
-                3. 大模型交流
-                4. 追加会话消息
-                5. 判断 是否调用工具
-                6. 是： tool 返回结果并追加，否： 追加并返回
-
-
-                Returns:
-
-                """
+        5. 判断 是否调用工具
+        6. 是： tool 返回结果并追加，否： 追加并返回
+        """
 
         # 5. 判断 是否调用工具
         if response_message.tool_calls:
@@ -106,3 +96,68 @@ class Engine:
         """把 LLM 响应转成 Message 。"""
         return response.message
 
+    """==========异步部分=========="""
+    async def ainvoke(self, input_message:Message) -> Message:
+        """但系调用llm，并存入记忆，无tool调用"""
+        self.conversation.append(input_message)
+        message = await self.achat()
+        return message
+
+
+    async def achat(self) -> Message:
+        """
+        1. 读会话消息
+        2. 大模型交流
+        3. 追加会话消息
+        """
+        # 1. 读会话消息
+        message = self.conversation.get_messages()
+
+        # 2. llm交互
+        response = await self.llm.ainvoke(self.messages_to_llm_request(message))
+
+        # 3. 追加会话消息
+        response_message = self.llm_response_to_message(response)
+        self.conversation.append(response_message)
+
+        return response_message
+
+    async def adeal_tool_or_output(self, response_message:Message) -> Message|None:
+        """
+        5. 判断 是否调用工具
+        6. 是： tool 返回结果并追加，否： 追加并返回
+        """
+
+        # 5. 判断 是否调用工具
+        if response_message.tool_calls:
+            # 6. 是： tool 返回结果追加
+            tool_results = await self.tool.ainvoke_many(response_message.tool_calls)
+            self.conversation.append_many([build_tool_message(tool_result.tool_call_id, tool_result.content) for tool_result in tool_results])
+            return None
+        else:
+            # 6. 否： 追加并返回
+            return response_message
+
+
+    async def arun(self, input_message:Message) ->Message:
+        """
+        1. 追加输入
+        2. 读会话消息
+        3. 大模型交流
+        4. 追加会话消息
+        5. 判断 是否调用工具
+        6. 是： tool 返回结果并追加，否： 追加并返回
+
+
+        Returns:
+
+        """
+        self.conversation.append(input_message)
+        while True:
+
+            response_message = await self.achat()
+            result = await self.adeal_tool_or_output(response_message)
+            if result is None:
+                continue
+            elif isinstance(result, Message):
+                return result